requirements.txt
asgiref==3.3.1
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
Django==2.2.17
django-cloudinary-storage==0.3.0
django-extra-views==0.13.0
django-haystack==3.0
django-localflavor==3.0.1
django-oscar==2.1
django-oscar-paypal==2.0.0
django-paypal==1.0.0
django-phonenumber-field==3.0.1
django-tables2==2.2.1
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
factory-boy==2.12.0
Faker==5.3.0
gunicorn==20.1.0
idna==2.10
phonenumbers==8.12.15
Pillow==8.1.0
psycopg2==2.8.6
purl==1.5
pycountry==20.7.3
python-dateutil==2.8.1
python-stdnum==1.14
pytz==2020.5
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
stripe==2.55.2
text-unidecode==1.3
urllib3==1.26.2
whitenoise==5.2.0
